chore: 기존 amplitude api 응답 처리 로직 삭제
related issue #585
